Refactor some pokemon util functions
