+++ src/me/corriekay/pokegoutil/utils/pokemon/PokemonCpUtils.java
@@ -9,6 +9,10 @@ package me.corriekay.pokegoutil.utils.pokemon;
 import java.util.HashMap;
 import java.util.Map;
 
+/**
+ * We don't need this class anymore, cause the Pogo-API can do all that his class does as well.
+ */
+@Deprecated
 public class PokemonCpUtils {
     private static final Map<Float, Float> LEVEL_CPMULTIPLIER = new HashMap<>();
 
@@ -147,10 +151,10 @@ public class PokemonCpUtils {
     /**
      * Get the CP for the given values
      *
-     * @param attack       All attack values combined
-     * @param defense      All defense values combined
-     * @param stamina      All stamina values combined
-     * @param level the Pokémon level for which the CP should be calculated
+     * @param attack  All attack values combined
+     * @param defense All defense values combined
+     * @param stamina All stamina values combined
+     * @param level   the Pokémon level for which the CP should be calculated
      * @return Maximum CP for these levels
      */
     public static int getCpForPokemonLevel(int attack, int defense, int stamina, float level) {
