+++ src/me/corriekay/pokegoutil/windows/PokemonGoMainWindow.java
@@ -22,6 +22,7 @@ import me.corriekay.pokegoutil.utils.ConfigNew;
 import me.corriekay.pokegoutil.utils.helpers.FileHelper;
 import me.corriekay.pokegoutil.utils.helpers.UIHelper;
 import me.corriekay.pokegoutil.utils.logging.ConsolePrintStream;
+import me.corriekay.pokegoutil.utils.pokemon.PokemonCalculationUtils;
 import me.corriekay.pokegoutil.utils.pokemon.PokemonUtils;
 import me.corriekay.pokegoutil.utils.ui.SmartScroller;
 import me.corriekay.pokegoutil.utils.version.Updater;
